modify : QuoteStack 분리
